% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddQuarter}
\alias{AddQuarter}
\title{Add a quarter to a date (in format YYYY-MM-01 with MM in {01, 04, 07, 10})}
\usage{
AddQuarter(currentdate)
}
\arguments{
\item{currentdate}{A date for the begining of a quarter}
}
\value{
A Date object
}
\description{
Add a quarter to a date (in format YYYY-MM-01 with MM in {01, 04, 07, 10})
}
