% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AddQuarterStr}
\alias{AddQuarterStr}
\title{Add a quarter to a date (in format AAAAQQ)}
\usage{
AddQuarterStr(datestring)
}
\arguments{
\item{datestring}{A date (string)}
}
\value{
A string containing the new date
}
\description{
Add a quarter to a date (in format AAAAQQ)
}
