% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dateMerge}
\alias{dateMerge}
\title{Merge two dataframes based on the date column}
\usage{
dateMerge(left, right)
}
\arguments{
\item{left}{a data.frame with a date column}

\item{right}{a data.frame with a date column}
}
\value{
A data.frame
}
\description{
Merge two dataframes based on the date column
}
