% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maximum_date}
\alias{maximum_date}
\title{Maximum of two dates}
\usage{
maximum_date(date1, date2)
}
\arguments{
\item{date1}{a Date}

\item{date2}{a Date}
}
\value{
max(date1, date2)
}
\description{
Maximum of two dates
}
