% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minimum_date}
\alias{minimum_date}
\title{Minimum of two dates}
\usage{
minimum_date(date1, date2)
}
\arguments{
\item{date1}{a Date}

\item{date2}{a Date}
}
\value{
min(date1, date2)
}
\description{
Minimum of two dates
}
