% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Qdate}
\alias{Qdate}
\title{Convert a date in format AAAAQQ in standard date format}
\usage{
Qdate(datestring)
}
\arguments{
\item{datestring}{a string containing the date in AAAAQQ format}
}
\value{
A Date object
}
\description{
Convert a date in format AAAAQQ in standard date format
}
