% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Qdatevect}
\alias{Qdatevect}
\title{Convert a date vector in format AAAAQQ in standard date format}
\usage{
Qdatevect(datevect)
}
\arguments{
\item{datevect}{a vector of strings containing the dates in AAAAQQ format}
}
\value{
A DateVector object
}
\description{
Convert a date vector in format AAAAQQ in standard date format
}
