% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recup_troll}
\alias{recup_troll}
\title{Read a Troll database in text format
The file is split by space character}
\usage{
recup_troll(str_vec)
}
\arguments{
\item{str_vec}{A vector of strings containing the data}
}
\value{
A List
}
\description{
Read a Troll database in text format
The file is split by space character
}
