% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recup_troll_rcpp}
\alias{recup_troll_rcpp}
\title{Read a Troll database in text format}
\usage{
recup_troll_rcpp(filename)
}
\arguments{
\item{filename}{A vector of strings containing the data}
}
\value{
A List
}
\description{
Read a Troll database in text format
}
