% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recup_troll_str_vec}
\alias{recup_troll_str_vec}
\title{Read a Troll database in text format
The file is split by space character}
\usage{
recup_troll_str_vec(filename)
}
\arguments{
\item{filename}{A string containing the name of the file}
}
\value{
A vector of strings
}
\description{
Read a Troll database in text format
The file is split by space character
}
