% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recup_trollts.R
\name{recup_trollts}
\alias{recup_trollts}
\title{Fonction de lecture des bases de données issues de Troll}
\usage{
recup_trollts(filename)
}
\arguments{
\item{filename}{nom du fichier .frm au format trolltxt (chaine de caractere)}
}
\value{

}
\description{
Fonction de lecture des bases de données issues de Troll
}
