% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_split_cpp}
\alias{str_split_cpp}
\title{function to split at spaces or eol (internal)}
\usage{
str_split_cpp(s)
}
\arguments{
\item{s}{A string containing the whole file}
}
\value{
A vector of strings
}
\description{
function to split at spaces or eol (internal)
}
